video-texture: Respect the pixel aspect ratio of the incoming frames
Video frames come with a pixel-aspect-ratio cap that tells the sink what
should be the final size of the video.

This information is now given to ClutterGstVideoTexture that derives the
preferred width and height based on the base size of the frames and that
pixel-aspect-ratio.

A new debug category "aspect-ratio" has been created to trace what
(cherry picked from commit 7dd2c2b2b78bd2146fff0f84b99b33c342632116)