sink: Make the sink hold a weak reference to the ClutterTexture
The sink does not need to hold a reference to the Texture, it better to
leave the life cycle of the ClutterTexture entirely to Clutter.

If the texture disapears when the sink is still live and pushing buffers
to it, you will get warnings not nothing dramatic. It's your
responsibility to ensure the GStreamer pipeline has been stopped when
removing/destroying the texture.
(cherry picked from commit ddbb091b4a4d4cce0f3b324819af975a246faf4d)