video-sink: keep GstBuffer around until the next frame
We're working around the blitter here...

The only supported GL texture upload plugin (as far as I know) is
vaapidecode at the moment. This plugin has a fairly slow upload path :

vaapisurface -> Pixmap -> GL texture

This can take some time and the last part is currently not
synchronized. Therefore it leads to tearing. This patch tries to keep
the buffer around to prevent that but it's not perfect.