sink: Always create new texture in dmabuf
COGL does not properly redraw when we simply bind a new EGLImage to
the texture, which lead to wrong rendering when nothing else is being
composed. This is the side effect of doing GL outside of COGL. To
workaround, always create a new texture handle when a frame is uploaded.
This is relatively cheap and ensure the COGL state get dirtied.